requirements.txt
# Web scraping and data collection
requests==2.31.0
selenium==4.15.2
beautifulsoup4==4.12.2
scrapy==2.11.0
fake-useragent==1.4.0

# Database
pymysql==1.1.0
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# NLP and AI
torch==2.1.1
transformers==4.35.2
accelerate==0.24.1
datasets==2.14.6
tokenizers==0.15.0

# Jupyter and development
jupyter==1.0.0
ipykernel==6.26.0
notebook==7.0.6

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
dateutils==0.6.12
python-dateutil==2.8.2

# Web framework (if needed for API)
fastapi==0.104.1
uvicorn==0.24.0

# Data validation
pydantic==2.5.0
